app, libgimp, pdb: fix other int sizes.
Avoid the stack smashing bug from yesterday in other functions.
Additionally to fixing other functions, do not cast the pointer to size
in the PDB generation scripts so that we can quickly spot such bugs in
the future, through compilation time warnings, instead of hiding them.

+++ libgimp/gimpdrawablecolor_pdb.c
@@ -207,7 +207,7 @@ gimp_drawable_colorize_hsl (GimpDrawable *drawable,
 gboolean
 gimp_drawable_curves_explicit (GimpDrawable         *drawable,
                                GimpHistogramChannel  channel,
-                               gint                  num_values,
+                               gsize                 num_values,
                                const gdouble        *values)
 {
   GimpValueArray *args;
@@ -260,7 +260,7 @@ gimp_drawable_curves_explicit (GimpDrawable         *drawable,
 gboolean
 gimp_drawable_curves_spline (GimpDrawable         *drawable,
                              GimpHistogramChannel  channel,
-                             gint                  num_points,
+                             gsize                 num_points,
                              const gdouble        *points)
 {
   GimpValueArray *args;
